Rev 0.0.1.3

+++ Application/simpliciti/Applications/application/End_Device/main_ED_BM.c
@@ -37,7 +37,8 @@
 #include "bsp_leds.h"
 #include "bsp_buttons.h"
 #include "simpliciti.h"
-#include "driver/display.h"
+//#include "driver/display.h"
+#include "display.h"
 #include "rfsimpliciti.h"
 
 
@@ -51,7 +52,7 @@
 #define CONV_MS_TO_TICKS(msec)         			(((msec) * 32768) / 1000) 
 
 // U16
-//typedef unsigned short u16;
+//typedef unsigned short unsigned int;
 
 // *************************************************************************************************
 // Prototypes section
@@ -62,7 +63,7 @@
 extern uint8_t sInit_done;
 
 // SimpliciTI has no low power delay function, so we have to use ours
-extern void Timer0_A4_Delay(u16 ticks);
+extern void Timer0_A4_Delay(unsigned int ticks);
 
 extern unsigned char simpliciti_payload_length;
 //extern txOpt_t  simpliciti_options;
